cleaning and adding tsx dependencies
