+++ frontend/package.json
@@ -16,7 +16,9 @@
     "expo-random": "~13.1.1",
     "expo-dev-client": "~2.2.1",
     "expo-web-browser": "~12.1.1",
-    "expo-splash-screen": "~0.18.2"
+    "expo-splash-screen": "~0.18.2",
+    "typescript": "^4.9.4",
+    "@types/react": "~18.0.27"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
